test(ws-client): shared AsyncDisposable startServer helper
