+++ packages/ws-client/tests/browser-transport.test.ts
@@ -1,28 +1,12 @@
-import { once } from 'node:events'
-import { createServer } from 'node:http'
-import type { AddressInfo } from 'node:net'
-// eslint-disable-next-line import/default
-import httpTerminator from 'http-terminator'
 import { WebSocket as UndiciWebSocket } from 'undici'
 import { describe, expect, it } from 'vitest'
-import type { WebSocket } from 'ws'
-import { WebSocketServer } from 'ws'
 import { WebSocketConnectionEngine } from '../src/connection.ts'
 import { BrowserTransport } from '../src/transport/browser-transport.ts'
 import type {
   TransportFactory,
   TransportHandlers,
 } from '../src/transport/transport.ts'
-
-async function startServer(onConnection: (ws: WebSocket) => void) {
-  const server = createServer()
-  const { terminate } = httpTerminator.createHttpTerminator({ server })
-  const wss = new WebSocketServer({ server })
-  wss.on('connection', onConnection)
-  await once(server.listen(0), 'listening')
-  const port = (server.address() as AddressInfo).port
-  return { url: `ws://localhost:${port}`, terminate }
-}
+import { startServer } from './_util/server.js'
 
 // Factory that binds the undici WebSocket as the browser-global stand-in.
 const factory: TransportFactory = (url, handlers, options) =>
@@ -30,9 +14,8 @@ const factory: TransportFactory = (url, handlers, options) =>
 
 describe('BrowserTransport via engine', () => {
   it('reports heartbeat:false, pauseResume:false', async () => {
-    const { url, terminate } = await startServer((ws) => ws.close(1000))
-    await using _ = { [Symbol.asyncDispose]: async () => terminate() }
-    const engine = new WebSocketConnectionEngine(factory, url)
+    await using server = await startServer((ws) => ws.close(1000))
+    const engine = new WebSocketConnectionEngine(factory, server.url)
     expect(engine.capabilities).toEqual({
       heartbeat: false,
       pauseResume: false,
@@ -43,12 +26,11 @@ describe('BrowserTransport via engine', () => {
   })
 
   it('yields text and ends on clean close', async () => {
-    const { url, terminate } = await startServer((ws) => {
+    await using server = await startServer((ws) => {
       ws.send('hi')
       ws.close(1000)
     })
-    await using _ = { [Symbol.asyncDispose]: async () => terminate() }
-    const engine = new WebSocketConnectionEngine<'text'>(factory, url, {
+    const engine = new WebSocketConnectionEngine<'text'>(factory, server.url, {
       dataMode: 'text',
     })
     const received: string[] = []
@@ -57,14 +39,17 @@ describe('BrowserTransport via engine', () => {
   })
 
   it('yields binary frames as Uint8Array', async () => {
-    const { url, terminate } = await startServer((ws) => {
+    await using server = await startServer((ws) => {
       ws.send(Buffer.from([9, 8, 7]))
       ws.close(1000)
     })
-    await using _ = { [Symbol.asyncDispose]: async () => terminate() }
-    const engine = new WebSocketConnectionEngine<'binary'>(factory, url, {
-      dataMode: 'binary',
-    })
+    const engine = new WebSocketConnectionEngine<'binary'>(
+      factory,
+      server.url,
+      {
+        dataMode: 'binary',
+      },
+    )
     const received: Uint8Array[] = []
     for await (const msg of engine) received.push(msg)
     expect(received).toHaveLength(1)
@@ -73,16 +58,15 @@ describe('BrowserTransport via engine', () => {
 
   it('send resolves (transport-accepted semantics)', async () => {
     const seen: string[] = []
-    const { url, terminate } = await startServer((ws) => {
+    await using server = await startServer((ws) => {
       ws.on('message', (d) => {
         seen.push(d.toString())
         ws.close(1000)
       })
     })
-    await using _ = { [Symbol.asyncDispose]: async () => terminate() }
     let onOpen!: () => void
     const opened = new Promise<void>((resolve) => (onOpen = resolve))
-    const engine = new WebSocketConnectionEngine<'text'>(factory, url, {
+    const engine = new WebSocketConnectionEngine<'text'>(factory, server.url, {
       dataMode: 'text',
       onOpen,
     })
